+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Populate building number in wayfinding building end points [#100](https://github.com/rokwire/gateway-building-block/issues/100)
+
+### Added
+- wayfinding/floorplans end point 
+
 [2.10.1] - 2024-05-22
 ### Fixed
 - Incorrect event end times [#97](https://github.com/rokwire/gateway-building-block/issues/97)